app: remove 'npm run check' step
For some reason, CircleCI was failing whenever it ran 'flow check' as
a subprocess of Make. The error ENOTBLK was returned by an fallocate
call, according to strace. It worked fine when running as a subprocess
of anything else, and none of the `--sharedmemory-*` flags to `flow
check` solved it.