read .sg-drone.yml instead of .drone.yml (#172)
Sourcegraph is currently using drone to run its builds. While this enables some reuse of drone functionality during the Sourcegraph build, our environment is still different to the normal drone environment. Using the same config file leads to conflicts that can not be easily resolved. This commit makes Sourcegraph use the .sg-drone.yml file instead, so the project maintainer can decide with configuration to share between the two environments and which parts to not run on Sourcegraph.
Resolves https://app.asana.com/0/91952577398025/133290561357955