remove Repos.List pagination
The pagination was never used in the UI (all repos were always shown),
and all other calls to Repos.List specified a very large PerPage
number to fetch all. So, it was totally unused.

Removing pagination (for now) simplifies the merging of Repos.List and
Repos.ListRemote, because it's hard to paginate data that results from
merging 2 underlying datasets.