remove Repos.ListRemote API, add Repos.List IncludeRemote option
The dual APIs for listing repos was adding complexity to any client
that needed to know if a remote repo already existed locally. The
combining of the 2 lists should occur on the server side, not be
pushed to the client, to centralize the merging logic and reduce the
amount of data that needs to be transferred.

This change only removes the gRPC API method, not the HTTP API or
JavaScript Flux code related to remote repos.