app: display "You must sign in to continue." on BetaInterestForm when displayed as modal
Prior to this change we did not properly display a sign in button when displaying
BetaInterestForm inside a modal, and this caused submission to fail with errors
about not having admin rights etc.

Upon sign in we redirect to `/beta`, for now, but this can be improved in the future.