services/backend/internal/localstore: Fix List RemoteOnly skipping local repos. (#244)
This is a followup to #230.

Fix issue in the logic of RemoteOnly handling that results in local
repos (ones that are present in DB) to be skipped from the list of
remote repos.

Populate repo ID (Sourcegraph ID from database) in returned repos even
when RemoteOnly if set. This isn't used now, but will be helpful later
once we rely more on repo IDs internally.

Add test.