all: fix submission of BetaInterestForm when account has no primary email address
When a GitHub user has their email address set to private, we do not get access
to it and thus their Sourcegraph account has no primary email address. Thus, we
need to ask them for it as part of the BetaInterestForm. To not bother users,
and also to avoid confusing situations, we only add the email address field if
the account has no primary email address (and it is subsequently set as the
primary account email address).

- Add and implement `Accounts.UpdateEmails` which supports adding email addresses
  currently.
- Implement and improve the `BetaRegistration.Email` field, such that it handles
  the logic described above exactly.