e2e: Add convenience function Click
We add the convenience function so that we can add a work-around on an issue
that affects ChromeDriver. Our assumption is that ChromeDriver will try to click
on an element, but between the time it gets the element and clicks on it, the
element moves. This happens a lot on the Usage Examples page, since we
asynchronously load in some extra annotations that appear above the `View all
references` button. This lead to internal errors from selenium, with selenium
logs like the following:

  org.openqa.selenium.WebDriverException: unknown error: Element is not clickable at point (810, 295). Other element would receive the click: <div class="Blob__margin___2uZsC" style="position: relative;">...</div>

More information at https://github.com/seleniumhq/selenium-google-code-issue-archive/issues/2766