github: Use an HTTP cache backed by Redis
Previously we used an in-memory cache for caching at the HTTP Transport layer in
the Github Client. This lead to large amount of uncached requests to the Github
API on startup (eg when deploying). This started to cause us to get
uncomfortably close to getting rate limited to no requests.

This change switches us from an in-memory cache to Redis. This will also have
the advantage of all instances sharing the same cache, further reducing the hits
we do against the GitHub API. A new type of Redis cache is introduced called
`rcache.ByteCache` which matches the interface used by `httpcache`.

Note that `httputil.Cache` and `httputil.CachingClient` is currently only used
by GitHub.

Fixes https://app.asana.com/0/87040567695724/154090028181702