app: link to the line of the ref when viewing an example
Previously, the "Used in REPO" link just linked to the repo homepage,
which was not very helpful.

This change makes it so that it now says "From REPO" and it links to
the file and line number that the ref came from. This is not perfect,
but it's a big improvement.