app: fix repeated WantFile and WantAnnotations actions
Previously the code was iterating over the list of definition results and issuing a WantFile and WantAnnotations for each of them, although they may all be in a few or even a single file. This means that the actions were issued multiple times for the same file. The new implementation iterates over the list of files instead.

Fixes https://app.asana.com/0/87040567695724/153775463609634