test/e2e: use a unique user account prefix for every e2etest user
Without this change it is possible for two users to hit the same target
Sourcegraph instance with e2etest and for one of them to fail. For example,
this causes some tests to be flaky when attempting to run against sourcegraph.com
because #e2etest is already running against it.

By using a username prefixed with an unique hardware ID, we can avoid this
issue all together.