test/e2e: split search_flow into one test per query
This requires some mild code duplication, but improves the run time
of this test (~30s avg for all -> ~9s avg for all). If this pattern
becomes troublesome in the future we can reverse it, but for now it's
a cheap win (esp in terms of running on CI, where total test time
matters most).