pkg/jsonrpc2: do not let SetResult and SetParams panic
Yes, in an ideal world they never should. But it still encourages bad error
handling and it will bite us in the future / make our services less reliable
(for example, it could take down the entire server if called not under the
presence of a panic handler, when perhaps 90% of requests would not cause a
panic but one does).