search: Limit concurrent refreshindex to 1 per repo (#299)
* queue: Allow a delay before running a job

This lets jobs decide to run in the future, but in a robust way. Our first use
will be RefreshIndex jobs trying again in the future if there is already an
update for the same repo.

* vendor github.com/hjr265/redsync.go

* rcache: A redis backed mutex

We want to be able to co-ordinate amongst different machines in the cluster. A
distributed mutex is better suited to something like consul/etcd/zookeeper, but
we already have redis in production and our initial uses will be coarse and
relatively long held locking around repo search updates. As such redis suits us
for the time being.

The API introduced is less than ideal, and can be improved. However, it should
suit our initial usecase well.

* rcache: Simplify test helper

* async: Ensure we only have 1 update to a repo at a time