services/backend/internal/localstore: fix GitHub auth for beta users / be less restrictive on account updates
Prior to this change, if a user was participating in a beta program (i.e. was already
granted access to one), the GitHub authentication flow would fail outright for them.
This was because the `Accounts.Update` method assumes that `len(modUser.Betas) > 0`
implies we are trying to update which betas a user is in, this had a few problems:

1. Updating betas is an admin-only operation, so it failed for normal users (and hence broke
   GitHub auth when the user was in a beta program).
2. It restricted us from ever fully removing a user from all beta programs.
3. It did not comparitively see if an update was in progress, for example if the existing user
   was already in the exact list of betas which we are to 'update' to (no real change).

To fix this, we query the existing user from the database to make the comparison. This lets us
check if they are actually, _truly_, trying to update their list of betas or not. Additionally,
it means that we can fully remove a user from all beta programs (which was not possible before).

This does add one more DB query to the authentication flow, but presumably it is still quite fast
in general. To improve this, we should consider not using an ORM / doing the queries directly
which would allow for just updating fields we care about updating -- but that is a much larger
task to be scoped outside this bug fix.

Confirmed fix locally against a bugged account using the following steps:

1. Sign in via GitHub
2. Grant access to a single beta via `src users update --set-betas="desktop" slimsag`.
3. Attempt to sign in again via GitHub -> on `master` this fails, with this change it does not.