pkg/langp: prevent potential data race between two workspace preparation requests
If one request is currently preparing a repository by e.g. cloning the git repository,
then the directory can exist while not being complete (i.e. still being cloned). The
second request would think the workspace is fully cloned when in reality it may only
partially be, which could lead to confusing and hard to debug situations.

Essentially this makes the contention around prepareWorkspace a `sync.Mutex` for each
unique repo@rev; but this is probably OK in terms of performance since most requests
will end up aborting early once they see that the workspace directory does exist and
is cloned. Otherwise we will need to implement a `sync.RWMutex` per repo@rev which is
much more complex and hard to reason about.