hotfix: don't include sourcegraph/sourcegraph in def.ref.locations results
This causes a problem when sourcegraph/sourcegraph contains the greatest number
of refs for a def key, but the current user does not have read permission for that repo.
The DefInfo page in this case shows no results due to bad pagination/filtering logic.
This won't solve the problem in all cases; it is just a temporary patch.