app: scroll on scrollable element (which is no longer the full window)
A previous change appears to have made only the "main content" element
scrollable, instead of the whole window. This meant that the previous
window.scrollTo call in Blob.tsx would be a no-op. We need to scroll
the element that is actually scrollable.

Fixes an issue where jump-to-def did not scroll to the target def,
either when you navigate to a def URL directly, jump to a def in
another file, or jump to a def in the same file.