services/repoupdater: add a ForceEnqueue method
This adds a way to bypass the enqueuement restrictions in place by the
Enqueue function, which are not suitable when the enqueuement is performed
only once (e.g. in the case of a githook). Also added is a prometheus counter
for measuring the number of forced enqueuements, and a gauge for measuring
the queue size.