all: Use go1.7 context
We do this for vendored directories as well. We do that, otherwise
`context.Canceled` and `context.DeadlineExceeded` will have two different
potential values, the ones vendored in and `pkg/context`. This will break code
were we check `ctx.Err()` outside of vendor, but it was canceled in a context
created in vendor.

The other possibility is to not update our use of context. However, when using
tools like `goimports`, it will constantly rewrite those. This approach instead
allows us to update dependencies to using `context`, as upstream supports it.

Also included in this change is switching CircleCI to go1.7.