app/internal/oauth2client: Pass modified context to linkAccountWithGitHub. (#781)
The http.Request context was being modified with
r = r.WithContext(sourcegraph.WithCredentials(r.Context(), ...), but then
linkAccountWithGitHub was being called with old ctx local variable.
That meant that the modification to the context wasn't propagated
forward to linkAccountWithGitHub.

Regarding style, it's not trivial to come up with the best pattern, should
we always use r.Context() everywhere instead of ctx? Or should we only
use ctx := r.Context() when there's no other way to get a context (i.e., it
wasn't passed in)? I've deferred a bigger style refactor until later, after
we figure out how to use context now that in 1.7 it's available inside the
http.Request.

Fixes #769.