services/backend/internal/localstore: Sync method names. (#782)
The method name in parameter to accesscontrol.VerifyUserSelfOrAdmin
did not match the actual method name. This change affects potential
error messages only and no other behavior, but it makes it easier to
find the relevant code if seeing the error message.

I ran into this when an error here was happening.