golang: Run gog per package
gog is ooming in production. The assumption is that gog keeps a lot of
information in memory, and passing every package to gog leads to that memory
growing too large. This tradeoffs speed with memory usage (gog has some constant
startup time + not re-using cached info between packages). However, that should
be small loss + gog is not directly user latency sensitive.

Fixes #798