pkg/langp: fix DefSpecToPosition and PositionToDefSpec
Previously the query was specified as `defspec-refs-internal` which returns
a list of references to the defspec, which is obviously incorrect. We would
then use the first one it returned -- which just happened to sometimes be
correct.

Using "exported" instead returns the location of the exported defspec, but
it does mean that DefSpecToPosition and PositionToDefSpec only work for
exported symbols now (which is fine long-term, because we only want def
keys for exported symbols anyway).