httpapi: use langp.Definition for j2d
Our current implementation of `PositionToDefKey` is very slow. It depends on
analysis having run, so using it for j2d is against one of the core goals of the
Universe project.

This commit will regress the behaviour were we load the margin on universe
repos. Filed https://github.com/sourcegraph/sourcegraph/issues/838 to track
that. It is essentially a revert of the httpapi changes in
https://github.com/sourcegraph/sourcegraph/pull/743