all: Refactor context style.
Follow the pattern of always using only r.Context() when an
*http.Request is available, and always using ctx when no *http.Request
is available.

Refactor handlerutil.Client to return only a *sourcegraph.Client and no
context. This requires the caller code to be more explicit about the
context it uses, rather than either using handlerutil.Client or
ignoring its context. This is more explicit, which is simpler and more
readable.

Use rcontext convenience package in
./services/httpapi.serveChannelListen handler to avoid a dangerous case
where there would otherwise have to be a local ctx variable in the
presence of a *http.Request.