pkg/jsonrpc2: make client/server symmetric per protocol spec
The previous jsonrpc2 package differentiated between the client and server. The
[JSON-RPC 2.0 spec](http://www.jsonrpc.org/specification), however, leaves open
the possibility that both sides of a connection can send and receive requests:

> The Client is defined as the origin of Request objects and the handler of Response objects.
> The Server is defined as the origin of Response objects and the handler of Request objects.
>
> One implementation of this specification could easily fill both of those roles, even at the same time, to other different clients or the same client. This specification does not address that layer of complexity.

This PR updates the jsonrpc2 package to be symmetric. LSP defines requests sent
by both the "client" and "server", so we technically need symmetry to support
all of LSP. I don't know if our usage of LSP requires the client to send any
requests, though. Even if not, this new impl of jsonrpc2 is much better and
eliminates race conditions and infinite blocking in case of network drops.

This implementation contains one known problem. A server handler doesn't have a
way to indicate it should close the connection. This leads to the connection on
the server never closing cleanly, and hence lots of log lines like

  jsonrpc2: protocol error: unexpected EOF