pkg/langp: Instrument *langp.Client requests to include request + response body/status code
This lets us see all requests to each language processor, and see exactly what their JSON
request body was, followed by the response from the language processor including status
code. We use `prefixSuffixSaver` which is a nicer human-readable version of `io.LimitReader`
used by the os/exec package for reading the JSON response, to prevent buffering large responses
in memory. We use a 1 KB limit for our traces because they are moderately high throughput
(the official recommendation is 4 KB for medium throughput tracing).

This is quite useful because it means you don't have to trail the kubernetes logs (which
wouldn't work after we have more traffic anyway) and can instead refer to the trace for this
information.