coverage: do not skew error reporting by requesting keywords we don't support
Prior to this change coverage would make one hover+j2d request for each keyword
that we don't support (like `nil`, `false`, etc) and `_` discard identifiers in
a file. We intentionally do not support these at the moment.

Comparing the tokens that coverage finds and requests:

| Description   | Tokens | Requests | Errors |
|------- -------|--------|----------|--------|
| Before change | 916    | 916      | 176    |
| After change  | 758    | 758      | 18     |

We can conclude that:

- (`916 - 758 == 158`) 158 tokens were for keywords that we intentionally don't support (`nil`, `false`, etc).
- 176 (errors before) - 18 (errors after) == 158 (tokens were for keywords).
- Our true error rate is: 18 errors out of 758 requests, or 2.37% of requests (these are because of #1004 #830 and #828).