golang: Include GenDecl for finding docs
`ast.TypeSpec` generally doesn't have documentation associated with it, but rather
the type decleration does (which is a `GenDecl`). We also keep looking further up
the scope until we find the information we want. This may cause us to associate
the wrong documentation with a point, but I'm not sure. Will refine this as we
find counter-examples.

Fixes #988