Monaco tweaks (#1034)
* ui: make monaco cursor an I-beam and not black

The black block cursor looked like a teletype terminal. I switched it
back to the I-beam, which looks nicer to me. To make it not feel like
a text insertion cursor, I made it gray (instead of black).

* ui: remove blob content placeholder on monaco view (unnecessary)

* ui: make tooltip z-index greater than global nav z-index to avoid tooltip clipping

Also removes the .raw.css special handling in Webpack. We can use
:global to set global CSS classes instead.

* ui: apply cursor commit to non-raw CSS file

* ui: eliminate double whitespace in hover tooltip

* ui: better monaco initialization and model handling

Uses less coarse componentWillReceiveProps instead of componentDidUpdate.

Fixes Firefox rendering issue where the height of the div was only
like 9px. Tested on both Firefox and Chrome.

* ui: re-layout monaco editor after window resize

Avoids making the bottom portion of the Monaco editor unreachable if
you resize the browser window.

* increase file size limit, prevent display of large files on pre-monaco blob view on frontend

* ui: hyphenate "right-click"

* ui: do not throw if there are no FmtStrings

* ui: do not navigate to dummy page on 404, use editor behavior (no action)

* only use flex column on monaco

* ui: allow backslash path separator for windows (monaco uses backslash on windows)

* ui: highlight lines when scrolled to

* ui: more reliable jump-to-def using React dirtyness, not globals

* fix lint

* npm run generate