services/backend: Improve style consistency in tests. (#1060)
There was some variety in style/order of setting up the context and
declaring a repos variable `s`. This commit attempts to apply the same
style everywhere.

It also removes a few places that attempted to do:

	ctx = store.WithRepos(ctx, &repos{})

That code initially confused me, because it was present in some tests
but not others. After investigating, I learned it was actually a noop.
It has no effect (in the situation where it's used). Remove it to
simplify code and improve readability.

These changes are backported from the original wip branch for #917.