universe: support the Go stdlib (#1074)
* langprocessor-go: do not attempt to update dependencies of stdlib

Because:

1. The stdlib is self-contained, has no external dependencies that must be fetched
   via `go get`.
2. Our implementation requires a `go list` which doesn't work on stdlib by default
   (without special casing).
3. It wastes resources for literally no reason.

After this change, github.com/golang/go properly prepares it's workspace under
universe and thus it isn't discarded after each use (which previously incurred
a ~30s or so delay between hover/j2d requests).

* lang/golang: special case 'go list' for the stdlib (github.com/golang/go)

This makes it such that `workspace/symbol` requests for the stdlib, e.g.
the ones that occur to list exported defs, successfully get to at least
the `gog` phase. They do fail at that point, because gog is disabled by
us intentionally, so this change has no real user impact at all -- it is
purely a correctness change for the future.