services/backend: Use authenticated GitHub client for repo search queries.
Do not use unauthenticated GitHub client at all. This resolves a
planned TODO.

The reason for this change is because using an unauthed GitHub client
in production is not viable. If 50 people do a repository search in the
same minute, 30 of them will get results and 20 will not, because the
rate limit is shared between them all.

Authed clients means each user has their own rate limit for doing
searches, so the rate limit is more realistic.

This also means that only users who have connected a GitHub account to
their Sourcegraph account will get this functionality.