universe: Do not 500 on hover/j2d of comments/string literals/etc (#1110)
* lang/golang: Empty Location for string literals/comments

Any position in a document is valid for vscode to send for hover or j2d. In the
case of a position for something without a definition (ie string literal, etc),
we can return an empty position.

This is currently the most common reason we 500 in the frontend. This commit
works towards us doing something more reasonable.

* universe: Support unresolved hover

We introduce a new boolean Hover field, `Unresolved`. If it is true, it means
the user hover was valid, but there is no information to display. Alternatives
considered:

* Empty Title: it would hide bugs where we can't find a title.
* HTTP status codes: none really fit, and I am personally a fan of encoding
  semantics into the response body rather than status codes.

* universe: noop hover/j2d on struct field initializers.

We explicitly do not support this feature yet. See
https://github.com/sourcegraph/sourcegraph/issues/1104