pkg/langp: fix btrfs support
This effectively fixes btrfs support for langprocessor-go and others
because the language processor does not (and should not) run with `SG_FEATURE_UNIVERSE=t`.
In fact, it was my mistake that I ever introduced a src-process feature flag into this
code.

Instead, we can rely on the binary name to check that we are a `src` binary
instead of the feature flag. This prevents all `src foobar` commands from
spewing a warning about btrfs not being available.