ext/github: Do not make GH API requests for public repos for auth'd users. (#1202)
Previously, we would unconditionally make a request to GH's API if
a user was authenticated. This is wasteful, since the only repos we
cache are already known to be public.

This change makes it so that authed users use cache in two ways.

One, if a repo is unavailable (404) even to an authed user, that 404
response is cached so that future unauthed users can return from
cache right away (authed users will still check, in case it's a private
repo that only they have access to).

Two, if a repo is available to an authenticated user and happens to
be public, it's added to the public repo cache (with the repo permissions
cleared to defaults). Then, future authed and unauthed requests for
that public repo will be served from cache. This means that an authed
user with higher permissions may see lower permissions, but that's okay
for now because we don't use repo permissions for anything.