pkg/gitserver: Refactor into server and client.
This makes the code IMO much more readable and easier to reason about.
That will enable changes to be done in future commits.

Previously, it was really hard to see which code was running on the
server, and which code was running on the client. It was hard to know
which global variables were being accessed/modified on the server, and
which ones were accessed/modified on the client. It was also hard to
remember which APIs are available on the server and which are available
on the client.

This change fixes all those issues.

Also be more consistent about internally naming the library "gitserver"
rather than "git-server" or "GitServer". This way, it can be referred
to as "gitserver server" and "gitserver client". But flags and env vars
remain unchanged for compatibility.