services/backend: Timebox certain ls-tree using operations.
Set a timeout on the request context for repoTree.{Get,List,Search} and
repos.GetInventory. Those seem to be the main endpoints that indirectly
perform ls-tree operations, sometimes recursively.

Add support for checking if context is cancelled in gitserver client
and do not make network calls in that case.

This should help catch and/or limit cases gitserver goes out of control
with ls-tree operations, since each original request has a limit on
time while it can issue additional gitserver client requests.

Resolves #1256.
Updates #1169.