repoupdater: access repo data directly from store
Making this change to prevent the repo updater from calling
the Repo.Get via the backend which ends up hitting the API. Since repo
updater runs often this counts against our rate limit. It shouldn't be
critical for the repo updater to have completely fresh metadata when it
runs so fetching from the DB instead should be fine.