store: Use langp to update global indexes (#1268)
* defs: Rename UpdateFromSrclibStore to Update

It will no longer be from the srclib-store.

* defs: If refs indexing fails, still attempt to do defs indexing

See added comment for why this is valid. In particular we may not have indexing
for refs working just yet in universe, but we will for defs. This will allow us
to index defs in the meantime.

* store: Index global usage and search from universe

We remove support for indexing from the srclib-store. As it is, this store is
not being updated.

* global_refs: Directly use langp.ExternalRefs for Update

The previous implementation was never used in production, and upon actual
testing did not work as expected. This should be the correct implementation, and
solely relies on what langp returns.

* refs: Disable update until langp.ExternalRefs is acceptable

* refs: Use langp.Ref instead of graph.Ref for update

We can directly use the data we get from langp and avoid one extra copy.

* store: Disable failing tests for global indexes

They are currently very tied to srclib, which make them a pain to update right
now. Filed https://github.com/sourcegraph/sourcegraph/issues/1276 to track
updating them.