langp: Use Repos.GetInventory to only send to relevant language processors (#1299)
* golang: Marshal empty lists correctly

`var x []Foo` is `nil` until appended to. So in the case of an empty response
(ie no exported symbols), we would marshal null. The jsonrpc2 library would then
return the following to the client

```
jsonrpc2: protocol error: jsonrpc2: unable to determine message type (request or response)
```

I just audited all our uses of this pattern in golang rather than doing a
generic fix. I spent a bit of time thinking of a more generic fix that didn't
hide mistakes, but couldn't think of something. This should do for now.

* langp: Use Repos.GetInventory to only send to relevant language processors

Previously we sent to every processor for requests scoped to the repo, rather
than file.

* httpapi: Add endpoint to manually trigger RefreshIndexes