landing page: cleanup where template data is coming
This makes it much more easy to see where the template data is coming
from. In the template, we do not typically need the large data structures
and using them makes migrating to universe harder. So, instead, only pass
in data that we actually need access to from within the template.