httptrace: Use parent trace if specified as header
This will allow us to get better traces for uses of the httpapi internally.
Currently that is just universe's access and github token calls, but will be
used in xlang as well soon.

To avoid tracing XHR requests, we had to avoid injecting the information into
the JSContext. I don't believe this is actually used anywhere.