services/backend: re-enable GetSrclibDataVersionForPath
This should not have been added as part of [PR 1432](https://github.com/sourcegraph/sourcegraph/pull/1423)
by me because it causes srclib to only consult the latest version of it's data.
We will want this once we move the def landing page off of srclib, but for now
it causes most def landing pages to 404 due to having no data for the most
recent commit since we turned off srclib builds.

Mitigates #1311