xlang: don't buffer github archive .zip files
If the GitHub git repo archive .zip file is already cached on disk, then
we can just open the file on disk and avoid buffering it all into
memory.

If we do so, we need to close the .zip file when we're done, so this
commit also closes the LSP server proxy's workspace file system when it
shuts down.