xlang: remove textDocument/didOpen, use VFS between LSP proxy and build/lang server
Instead of sending all files in textDocument/didOpen requests, the
build/lang server now access the workspace file system by sending
fs/readFile, fs/readDir, fs/stat, etc., requests to the LSP proxy.
They only do so for the files that are actually needed.

This greatly improves performance. For example, the Kubernetes repo
now (when the .zip archive is cached on disk) now can return a "hover"
result in 2.7s, as opposed to 13s previously. The speedup is due to
not having to send all of the file contents to the build server; only
the specific files needed are requested by the build server.