xlang: never try to fetch+overlay the current workspace's repo during go deps fetch
Previously, if a repo imported something underneath its own import path
that didn't exist, or if any of its dependencies did the same to it, the
Go dep fetcher would try to re-fetch the same repo and overlay it on top
of itself. This wasted time and could cause weird bugs.