xlang: suppress invalid defs for builtins (fixes #1474)
Previously, the error `invalid relative file path in URI` was showing
up whenever a definition lookup yielded a types.Object with an invalid
pos (obj.Pos.IsValid() == false). Suppress those for now instead of
returning an error message; they usually refer to things that it's not
crucial to be able to j2d for.