xlang: in "src xlang sendx?" commands, fetch code from codeload for easier prod debugging
This way, you don't have to first clone in the UI to repro an issue
occurring on a public repo.

Downsides: adds another dev != prod disparity, doesn't work for private repos.

Let's see if this is better over the next 24h. The alternative is to
remove the "src xlang send" command altogether and rely on people
curling.