app/internal/ui: rm landing_test.go
The two tests in this file both rely heavily on mocking in order to test, and
this causes both of them to nil pointer deref with the new implementation. It
is my stance that these tests do not actually test very much, but rather just
confirm that 'we did call some functions'. One could very much so NOT rely on
these tests to confirm that the page actually works, it must be tested manually,
and thus they just are not good tests. The require updating often, are hard to
debug/fix (I've had to several times in the past), and are just all around not
as good as having proper unit tests would be.

Remove them for now, with the intent of adding proper unit tests in the future.