ui: fix find definition for last character of word issue
The range of a word includes the character after the word. This caused
an issue where if you hover over the character after a word, it would
not fetch the correct definition for that word since the position is
outside of the word, but this.currentWordUnderMouse would be updated
which would cause hovers over preceding characters in that word to not
trigger a fetch.