pkg/gitserver: Add tracing, log details of timed out requests. (#1546)
This way, we can find more information about the kinds of requests that
end up taking too long and get cancelled. This includes the repository
for which it happens, and the operation type (and its flags/options).

Trace all errors in broadcastCall, not just context deadline exceeded
errors. Should be able to filter/find the deadline exceeded ones via
"err" span tag.