xlang/golang: Granular locking for readFile
Reading from the FS can be blocking if it needs to go out to the network.
However, the lock we acquire is to protect running `h.reset`, not to synchronize
access to the VFS. The underlying FS's have there own synchronization, except
the overlayFS, so we introduce a lock for `overlayFS`.

I noticed this due to some operations being blocked which shouldn't have been.
So I audited every call to `.Lock` in `xlang/golang`. This seems like the only
place were our lock has underlying blocking operations. However, I do not think
it is the cause of the original issue I spotted.