xlang: don't acquire proxy mutex when fwding diagnostics to clients
Previously, if a client's connection blocked writes (in jsonrpc2.go
line 309 c.w.Flush()), it would block the server's forwarding of
diagnostics for all clients, all the while holding the proxy's mu.
This would prevent new proxy conns from being started (or from being
idle-killed).

This caused downtime on 2016-10-06 around 13:45 PDT. The cause is not
yet verified, but it seems likely to be this (and the second-order
cause is the complexity of the sync/mutex code in the LSP proxy).
Either way, this change is worthwhile. It is also still unknown what
traffic pattern from the HTTP API led to this issue.