ui: add tsconfig.lint.sh
We're importing vscode, which is not ready for non-nullable types, but we
want to do strict null checking in out codebase. Workaround this by running
the type checker using a config file that does not do strict null checks
first. Then to another type check with strict null checks turned on but skip
*.d.ts files, which are the source of the null issues in the vscode
project.