{landing page, global refs}: only show max 3 repos, 5 files for external refs
This mitigates an issue where pages like:

- https://sourcegraph.com/github.com/golang/go/-/info/GoPackage/fmt/-/Printf

Would show 3 repositories and a staggering 988 "referenced in XYZ.go" lines.
Not only is this bad from a UX perspective, it also caused the page to take ~50s
to load due to the number of references which must be checked for auth and
correctness.