pkg/handlerutil: Document GetRepoID.
It was not documented at all before. This made it hard to understand
what its behavior is supposed to be. In turn, it made it hard to change
behavior of other code that uses it.

The goal of this documentation is to describe current behavior, there
shouldn't be any changes in behavior. That would be a fault in this
commit.