services/backend/internal/localstore: Fix incorrect method.
Its value is not used at this time, so this is a noop. But fix it since
I spotted it, for consistency, and to avoid other people in the future
wasting time confirming if this has negative behavior implications (it
doesn't, because it's unused) and can be safely fixed (it can be).