services/backend/{accesscontrol,internal/localstore}: Use VerifyActorHasGitHubRepoAccess correctly.
The VerifyActorHasGitHubRepoAccess method is documented to strictly
require repoURL to have a "github.com/" prefix (case sensitive). That
was the assumption on which the change in behavior in PR #1760 was
done. This is a followup that corrects deviations from that behavior in
calling code.

It does so by checking that the repoURL has a "github.com/" prefix
(case sensitive) or a casing variation thereof (which is rejected as
bad input).

These are the only 4 places in the codebase that call
VerifyActorHasGitHubRepoAccess, and all 4 of them are augmented in the
same way with this check.

Optionally, we can factor out this check into a common new function to
reduce duplication from 4 to 1, at the cost of an extra level of
calling depth. Alternatively, we can change the API of
VerifyActorHasGitHubRepoAccess to have behavior that it had prior to
change in PR #1760 (which did not match its documentation, as Richard
and I understood it at the time) to accept all repoURIs and only
perform a check for ones that are considered to be github.com
repositories. Then this commit can be reverted.