New homepage (#1822)
* Implement top section of the homepage

* Add placeholder for feature carousel

- Get alignment and responsiveness right

* Add transitions to the carousel. Abstract carousel to it's own component.

* Fix tests, use better typing, ad sprinkles

* Align sprinkles, add language section

* Refine alignment and positioning

* Make language support accurate to NOW. Make sure beta link works.

* Replace SVGs with PNGs

* Fix SVG background rendering on retina

* Clean up code and types. Remove home.css. Compress images.

* Add typescript logo

Fixes #1829

* Update copy

Fixes #1832

* Fix whitespace issues

* Make sure footer wraps correctly on narrow screens and mobile

Fixes #1838

* Add TypeScript, correct capitalization of JavaScript

Fixes #1848

* Use react router Link for appropriate links in footer, update copy, add better typing to Logo

* Fix momentum scrolling bug and tweet scrolling bug

Fixes #1837

* Use latest version of glamor.

Includes the fix for: https://github.com/threepointone/glamor/issues/58

* Refactor to use initial values and less ternarys

The ternary operations on the media breakpoints ensure that column is
never looking for a “null” index, which would cause a JS error. There’s
not a good default here and the alternative would be to refactor the
entire thing into an overly complex if statement. I’m sure there’s a
more elegant solution, but I’m punting on it for now.