ui: Disable workspace/symbols for typescript
This is currently impacting the performance other core LSP actions (such as
hover and j2d) since it is CPU/IO intensive. We are working on a fix, but in the
meantime it probably makes sense to disable symbol support so more core features
work.

This commit is a hack without me understanding a better way to do this. Please
take over this PR / revert / etc if this is not a decent workaround.

See https://github.com/sourcegraph/sourcegraph/issues/1828 and
https://github.com/sourcegraph/sourcegraph/issues/1860